+++ release_notes.md
@@ -20,4 +20,5 @@
 
 ### 🐛 Fixed Issues
 
-- Fixed OData V2 error: Disable Validation in the absense of DOCTYPE
+- Fix OData V2 error: Disable Validation in the absense of DOCTYPE
+- Fix OData V2/V4 error: Navigating `nextLink` in paginated result-set no longer results in duplicate query parameters.